Add TODO plugin service for HTTP
    Add a plugin for managing a TODO list with service for HTTP protocol. Features of this plugin are: add, remove, and viewing of TODOs for a user by using their username. The corresponding endpoints are `GET /todos/<username>`, `POST /todos/<username>`, and `DELETE /todos/<username>`. This commit includes the manifest, OpenAPI spec, and the main Rust application file. The plugin logo is also included in the repository.

Updates described in this commit are an addition to HTTP querying to the plugin to interact with TODO lists. The server can be accessed by using the domain `http://localhost:8000`. This plugin was implemented with CORS restrictions configured to allow connections from `localhost:5002`, `localhost:8000`, `127.0.0.1:8000`, and `https://chat.openai.com`.

In addition, a `healthcheck` endpoint was added to simply respond with "I'm a todo plugin for chatgpt!".

No tests have been added so far.

    ~~~~~~~~~~
    This commit message was generated by HappyCommit. Try it in your project today!
    Check it out at https://github.com/jackbackes/happycommit